bifurcating components specific to bootstrap out - see github.com/heucuva/europi-bootstrap - this slims down the firmware-specific code significantly
