--- bootstrap/nonpico_panic.go
@@ -1,8 +1,0 @@
-//go:build !pico
-// +build !pico
-
-package bootstrap
-
-func init() {
-	DefaultPanicHandler = handlePanicLogger
-}
