--- bootstrap/nonpico_websimenabled.go
@@ -1,8 +1,0 @@
-//go:build !pico && websim
-// +build !pico,websim
-
-package bootstrap
-
-func init() {
-	defaultWebSimEnabled = true
-}
