--- bootstrap/pico_panicenabled.go
@@ -1,8 +1,0 @@
-//go:build pico && onscreenpanic
-// +build pico,onscreenpanic
-
-package bootstrap
-
-func init() {
-	DefaultPanicHandler = handlePanicOnScreenLog
-}
