--- experimental/knobmenu/item.go
@@ -1,10 +1,0 @@
-package knobmenu
-
-import "github.com/awonak/EuroPiGo/units"
-
-type item struct {
-	name     string
-	label    string
-	stringFn func() string
-	updateFn func(value units.CV)
-}
