--- internal/projects/clockgenerator/module/config.go
@@ -1,14 +1,0 @@
-package module
-
-import "time"
-
-const (
-	DefaultGateDuration = time.Millisecond * 100
-)
-
-type Config struct {
-	BPM          float32
-	GateDuration time.Duration
-	Enabled      bool
-	ClockOut     func(high bool)
-}
