+++ internal/projects/diagnostics/diagnostics.go
@@ -9,7 +9,6 @@ import (
 	"tinygo.org/x/tinyfont/proggy"
 
 	europi "github.com/awonak/EuroPiGo"
-	"github.com/awonak/EuroPiGo/bootstrap"
 	"github.com/awonak/EuroPiGo/experimental/draw"
 	"github.com/awonak/EuroPiGo/experimental/fontwriter"
 )
@@ -95,7 +94,7 @@ func main() {
 	}
 
 	// since we're not using a full bootstrap, manually activate the webservice (this is a no-op on pico)
-	if ws := bootstrap.ActivateNonPicoWS(e.Context(), e); ws != nil {
+	if ws := europi.ActivateNonPicoWS(e.Context(), e); ws != nil {
 		defer func() {
 			_ = ws.Shutdown()
 		}()
