--- internal/projects/randomskips/module/config.go
@@ -1,6 +1,0 @@
-package module
-
-type Config struct {
-	Gate   func(high bool)
-	Chance float32
-}
